fix build error by dynamically importing pixi
