+++ package.json
@@ -1,6 +1,7 @@
 {
 	"name": "symbol-bucket",
 	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "svelte-kit dev",
 		"build": "svelte-kit build",
@@ -13,6 +14,7 @@
 		"format": "prettier --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
+		"@sveltejs/adapter-auto": "^1.0.0-next.55",
 		"@sveltejs/adapter-node": "next",
 		"@sveltejs/kit": "next",
 		"@tailwindcss/typography": "^0.5.2",
@@ -30,7 +32,6 @@
 		"tailwindcss": "^3.1.4",
 		"typescript": "^4.7.2"
 	},
-	"type": "module",
 	"dependencies": {
 		"@supabase/supabase-js": "^1.35.3",
 		"@zerodevx/svelte-toast": "^0.7.2",
